// front/services/models/log.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Global Testing
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { LogStep } from './log-step';

/**
 * 
 * @export
 * @interface Log
 */
export interface Log {
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'idExecution'?: string;
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'idScenario'?: string;
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'idStep'?: string;
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'humanDescription'?: string;
    /**
     * 
     * @type {LogStep}
     * @memberof Log
     */
    'step'?: LogStep | null;
    /**
     * 
     * @type {string}
     * @memberof Log
     */
    'createdAt'?: string;
}